Moved Databricks CREATE FUNCTION logic into Databricks dialect. Add CTE as function definition optio and add test case.
