+++ src/sqlfluff/dialects/dialect_sparksql_keywords.py
@@ -318,6 +318,4 @@ UNRESERVED_KEYWORDS = [
     "TEXT",
     "CHOICES",
     "REMOVE",
-    "PYTHON",
-    "SQL",
 ]
